Fixed compile warning.
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@113296 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/proc_open.c
@@ -216,7 +216,7 @@ PHP_FUNCTION(proc_get_status)
 	array_init(return_value);
 
 	add_assoc_string(return_value, "command", proc->command, 1);
-	add_assoc_long(return_value, "pid", proc->child);
+	add_assoc_long(return_value, "pid", (long) proc->child);
 	
 #ifdef PHP_WIN32
